DuneClient as composition of Query and Execution APIs
