+++ src/api/client.ts
@@ -2,9 +2,10 @@ import { DuneError, ResultsResponse, ExecutionState, QueryParameter } from "../t
 import { ageInHours, sleep } from "../utils";
 import log from "loglevel";
 import { logPrefix } from "../utils";
-import { ExecutionClient } from "./execution";
+import { ExecutionAPI } from "./execution";
 import { POLL_FREQUENCY_SECONDS, THREE_MONTHS_IN_HOURS } from "../constants";
-import { ExecutionParams, ExecutionPerformance } from "../types/requestPayload";
+import { ExecutionParams } from "../types/requestPayload";
+import { QueryAPI } from "./query";
 
 const TERMINAL_STATES = [
   ExecutionState.CANCELLED,
@@ -12,7 +13,15 @@ const TERMINAL_STATES = [
   ExecutionState.FAILED,
 ];
 
-export class ExtendedClient extends ExecutionClient {
+export class DuneClient {
+  exec: ExecutionAPI;
+  query: QueryAPI;
+
+  constructor(apiKey: string) {
+    this.exec = new ExecutionAPI(apiKey);
+    this.query = new QueryAPI(apiKey);
+  }
+
   async runQuery(
     queryID: number,
     params?: ExecutionParams,
@@ -24,18 +33,18 @@ export class ExtendedClient extends ExecutionClient {
         params,
       )}`,
     );
-    const { execution_id: jobID } = await this.executeQuery(queryID, params);
-    let { state } = await this.getExecutionStatus(jobID);
+    const { execution_id: jobID } = await this.exec.executeQuery(queryID, params);
+    let { state } = await this.exec.getExecutionStatus(jobID);
     while (!TERMINAL_STATES.includes(state)) {
       log.info(
         logPrefix,
         `waiting for query execution ${jobID} to complete: current state ${state}`,
       );
       await sleep(pingFrequency);
-      state = (await this.getExecutionStatus(jobID)).state;
+      state = (await this.exec.getExecutionStatus(jobID)).state;
     }
     if (state === ExecutionState.COMPLETED) {
-      return this.getExecutionResults(jobID);
+      return this.exec.getExecutionResults(jobID);
     } else {
       const message = `refresh (execution ${jobID}) yields incomplete terminal state ${state}`;
       // TODO - log the error in constructor
@@ -43,22 +52,27 @@ export class ExtendedClient extends ExecutionClient {
       throw new DuneError(message);
     }
   }
-
+  /**
+   * Goes a bit beyond the internal call which returns that last execution results.
+   * Here contains additional logic to refresh the results if they are too old.
+   * @param queryId - query to get results of.
+   * @param parameters - parameters for which they were called.
+   * @param maxAgeHours - oldest acceptable results (if expired results are refreshed)
+   * @returns Latest execution results for the given parameters.
+   */
   async getLatestResult(
     queryId: number,
     parameters?: QueryParameter[],
     maxAgeHours: number = THREE_MONTHS_IN_HOURS,
   ): Promise<ResultsResponse> {
-    let results = await this._get<ResultsResponse>(`query/${queryId}/results`, {
-      query_parameters: parameters ? parameters : [],
-    });
+    let results = await this.exec.getLastExecutionResults(queryId, parameters);
     const lastRun: Date = results.execution_ended_at!;
     if (lastRun !== undefined && ageInHours(lastRun) > maxAgeHours) {
       log.info(
         logPrefix,
         `results (from ${lastRun}) older than ${maxAgeHours} hours, re-running query.`,
       );
-      results = await this.runQuery(queryId, {query_parameters: parameters});
+      results = await this.runQuery(queryId, { query_parameters: parameters });
     }
     return results;
   }
@@ -71,6 +85,6 @@ export class ExtendedClient extends ExecutionClient {
     parameters?: QueryParameter[],
     pingFrequency: number = 1,
   ): Promise<ResultsResponse> {
-    return this.runQuery(queryID, {query_parameters: parameters}, pingFrequency);
+    return this.runQuery(queryID, { query_parameters: parameters }, pingFrequency);
   }
 }
