+++ src/api/index.ts
@@ -1,4 +1,4 @@
+export * from "./client";
 export * from "./execution";
-export * from "./extensions";
 export * from "./query";
 export * from "./router";
