+++ src/index.ts
@@ -1,3 +1,2 @@
-export { ExtendedClient as DuneClient } from "./api/extensions";
-export * from "./api";
+export { DuneClient, QueryAPI, ExecutionAPI } from "./api";
 export * from "./types";
